// src/models/dtos/aibook.dto.ts
export class AiBookDto {
    aiBookDataId: string;
    startPage: string;
    endPage: string;
    s3VideoURL: string;
    s3GenSummaryURL: string;
    statusId: string;
    status: string;
    bookId: string;
    bookName: string;
    teacherId: string;
    teacherName: string;
    subjectId: string;
    subjectName: string;
    classId: string;
    className: string;
    instituteId: string;
    instituteName: string;
    stateId: string;
    stateName: string;
    createDate: string;
}